multivendor_ecommerce/requirements.txt
django
postgreSQL
PgAdmin4
psycopg2-binary==2.9.3
sqlparse==0.4.2
djangorestframework
djangorestframework-simplejwt
httpie